로그인 액티비티 생성 및 Fragment들 한 파일로 정리함
