+++ android/app/src/main/AndroidManifest.xml
@@ -16,18 +16,19 @@
         tools:targetApi="31">
         <activity
             android:name=".login.LoginActivity"
-            android:exported="false"
-            android:label="@string/title_activity_login"
-            android:theme="@style/Theme.Android" />
-        <activity
-            android:name=".MainActivity"
             android:exported="true"
-            android:theme="@style/Theme.Android">
+            android:label="@string/title_activity_login"
+            android:theme="@style/Theme.Android" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity
+            android:name=".MainActivity"
+            android:exported="false"
+            android:theme="@style/Theme.Android">
         </activity>
     </application>
