+++ android/app/src/main/java/com/example/front/presentation/Iot/Iot_page01.kt
@@ -1,0 +1,23 @@
+package com.example.front.presentation.Iot
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.front.presentation.R
+
+class Iot_page01 : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_iot_page01, container, false)
+    }
+
+}
