+++ android/app/src/main/java/com/example/front/presentation/Iot/Iot_page02.kt
@@ -1,0 +1,22 @@
+package com.example.front.presentation.Iot
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.front.presentation.R
+
+class Iot_page02 : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_iot_page02, container, false)
+    }
+}
